# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Target Discovery Platform API
  version: 0.1.0
  description: Frontend-first contract for mock and future backend integration.
servers:
  - url: http://localhost:3000
paths:
...
  /api/references/{id}:
    get:
      summary: Get one evidence detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reference detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferenceDetail"
        "404":
          description: Not found
components:
  schemas:
...
    ReferenceDetail:
      type: object
      required: [id, title, abstract, keyFindings]
      properties:
        id:
          type: string
        title:
          type: string
        abstract:
          type: string
        keyFindings:
          type: array
          items:
            type: string
...